ci: Add bootstrap AppMap analysis workflow
Create a branch called appmap-analysis.
Commit these changes to the the appmap-analysis branch and open a Draft PR.

After the build runs to completion, verify that the AppMap Analysis build
artifact has been created and stored.

Then download the install patch file that's
in the Action summary. Apply the patch to the repo with:

$ cat patch | git apply

Then move on to the next commit in the setup process.